Created "tools"
The goal is to do file validation using JSON Schema's based on the KM Shared Library validation cases. This is much easier than spawning an instance of the database and application to validate the files.

The schema is more specific (restrictive) than the app validator. This shouldn't matter for the most part.